+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+* Adds support for retrieving crimes within a custom area
+* Retries rate-limited API requests using the response's `retry_after` value
+
 ## [8.0.0] - 2026-09-15
 * Removes support for Node.js 20.x
 * Adds support for Node.js 24.x